Add simple client for API communication
This commit adds a  pretty simple API client to handle communication
between the Ribose API server and the client, this is more likely a
wrapper on top of the `Net::HTTP` standard library.

In the upcoming commits we will tweak this to support more http verb
and also to handle API response and errors with full flexibility.