Retrieve the list of user's settings
This commit adds the interface to list all of the user's settings
It also adds a support helper which we will use to stub all of the
API interaction between Ribose server and our test suite.

Stubbing API request in the test environment will save use lots of
time but we might also need to do some actual API request, So later
we will also implement some sort of acceptance test that will test
the actual API interaction and that way it ensure everything works
as expected. Usage:

```ruby
Ribose::Setting.all
```