Use Sawyer for HTTP communications
Based on our API response structure, we have decided to go with the
`Sawyer` gem. It should work with our current response structure &
as well as with any ongoing changes.

Sawyer is based on Fraday, so we can customize it using some sort
of middleware. We have also extracted the API communication to an
internal `Ribose::Request` class, so this will be only place of
change in the future.