Add interface to retrieve a space details
This commit adds the interface to retrieve the details for a specific
space, and this interface will return the `Sawyer::Resource`, which
can be access as instance method or using `[]`.

This commit also extract the `fetch` action to a separate module, so
from now on we only need to include that module and everything else
should work as usual. Usage:

```ruby
Ribose::Space.fetch(space_id)
```