Add interface to create a new user space
Ribose discussions are organized based on spaces, and normally one
user will be associated with one or more spaces and they can also
create a new space when necessary.

This commit adds the interface to create a new user space, and it
is pretty simple, all we need to do is provide the details and it
will then return the newly created space on success.

```ruby
Ribose::Space.create(
  access: "private",
  space_category_id: 12,
  name: "The amazing Ribose Space",
  description: "Description about your space"
)
```