Remove `create` action related duplications
Pretty much most of our resource creation are duplicating the same
behavior, in a high-level the main differences is in attributes &
we should be able to share the remaining between other resources

This commit extract the `create` action to a separate module and
it also adds couple of hook method to provide a resource specific
validations. This commit also remove all of the existing create
related duplications.