Add interface to create a new file upload
The new Ribose API has added support for direct file upload to
AWS, and that requires us to follow a couple of steps steps to
successfully upload a file and then complete the process.

This commit adds an easier interface to upload a file using a
very simple structure and on the under neath it does all the
work to prepare a upload, and then do the actual upload to s3
and finally notify the Ribose API. Usages:

```ruby
Ribose::SpaceFile.create(
  space_id,
  file: "The complete file path",
  tag_list: "here is, some, tag, sample",
  description: "The description for the upload",
)
```

Note: This new upload API is not available to public yet.