Interface to create connection invitation
This commit adds an interface to create connection invitation
using this Ruby client. This interface expect `emails` and the
`body` as an arguments and it will send out the invitation to
that email.

One another interesting thing, the `emails` key also accepts
an array of emails and if we provide multiple emails with it
then it will create some mass invitations.

```ruby
Ribose::ConnectionInvitation.create(
  emails: ["email-one@example.com", "email-two@example.com"],
  body: "This contains the details message about the invitation",
)
```