Interface to reject space join request
One Ribose user might sent out a request to join a space from one
of the other user, but that user will decide if he will accept or
decline that request. This commit adds the interface to that will
reject the join space request.

```ruby
Ribose::JoinSpaceRequest.reject(invitation_id)
```