Add interface to update join space request
In the previous commits, we have added the `accept` and `reject`
interface for join space request, but on the underneath those are
actually invoking the `update` method with specified `state`.

This commit expose that internal `update` interface, so user can
use this one to update some details for any of the existing join
space request.

```ruby
Ribose::JoinSpaceRequest.update(
  invitation_id, new_allowed_attributes_hash,
)
```