Extract out resource related helpers
In the current version, we have extracted out multiple actions
and each of them are dependent on some sort of helper methods
to build the appropriate resource, resource path, everything
related to do a successful request and there are a couple of
duplications between those helpers.

And even worst, we have some inconsistency between those name
and their behavior, which might make it quite complex in the
future.

This commit extracted out those helpers to a new, dedicated
module, so everything in one place and it also cleanup the
existing duplications. This should also keep a consistency