Add interface to active a user account
In the previous commit, we have added the sign up request sending
out interface, and the next step is to activate that user account
and that's when this commit comes in.

This commit adds the interface to activate the user account using
the information user will received through the Email.

This commit also modify some of the existing classes to allow a
user to send a request without any authentication header in it

```ruby
Ribose::User.activate(
  email: "user@example.com",
  password: "ASecureUserPassword",
  otp: "OTP Recived via the Email",
)
```