Add Ribose flavored custom exceptions
Currently for the invalid response, the API client does not raise
any error but returns the response with the invalid status code.

This commit improves it a bit, and add the error handling middle
ware that will check the response and ensures that we raise the
correct error message based on the response.