Add support to create a new user session
Currently, the way our API client is work to have the client find
the API keys using the web application and then configure it here

But now that we have decided to enhance this user experience so
this commit will allow the user to create a session and that will
automatically do all the underlying work to retrieve the detail.

In the long run, we can also add the support to allow the user to
login and then configure this client automatically so they don't
have to do any of those manually. Usages:

```ruby
Ribose::Session.create(username: username, password: password)
```

Fixes: #75