Add support to allow multiple user client
This commit extend the `Ribose::Request` functionality and adds
the support for `client` option with all of it's interfaces.

If the users provide a `:client` option with any of the valid
action call then it will use that client configuration and if
not then it will use the default configuration.

All of our actual interfaces might not have support for the
option yet but we will add those in the future commits.

```ruby
client = Ribose::Client.new(user_data)
Ribose::Request.get(endpoint, client: client, **attributes)
```
Related: 46