Extract out the delete action to module
In the current codebase, there are couple of interfaces that
have been re-implementing delete again and again, which is bit
of redundant and it also has a sister concern `cancel`.

In the long term we should probably stick to one, but for now
let's extract that out to a separate module and also sets the
`cancel` as an alias.