WIP: Add interface for wiki increments
Editing the Note from anywhere other than the web interface is
something we've never considered before, but is something we find
useful in user acceptance test setup.  There are some parameters
that are actually required but provided implicitly in the web UI,
which are missing from the API doc at the time of writing.

We're now reconsidering the way this endpoint works and once we
have something finalized then maybe we can get back to this one

Ref: Sir. Jeffery :)