Add interface to retrieve a file details
This commit adds the interface to retrieve the details for a file
in any given user space. This interface expects us to provide the
`space_id` and `file_id` as required parameter and it also allow
us to pass an additional `option` hash with additional details.

```ruby
Ribose::SpaceFile.fetch(space_id, file_id, options = {})
```