Add support to retrieve calendar events
This commit adds the support to retrieve the list of calendars
events, and this endpoint also support multiple ids at the same
time. In the underneath it usages the `id_pack` gem and prepare
calendar ids, and it also usages the current data as `start` &
default length of 7 days. Usages:

```ruby
Ribose::Calendar.fetch(calendar_ids, length: 7, start: Date.today)
```

Fixes: #97