Add interface to list calendar events
This commit adds the interface to retrieve the list of calendar
events for any of the specific calendar. In the underneath, it
delegate the listing behavior to `Ribose::Calendar`, so it will
actually allow to retrieve events for multiple calendars at a
same time. Usages:

```ruby
Ribose::Event.all(calendar_id, option_hash)
```