Fix file upload related error
On the file upload endpoint, the final `notify` returns the
content header as `text/html` which causing an issue with the
`Sawyer` gem, as it expects the response header to be `json`
otherwise it won't parse.

This commit adds a temporary fix, that will bypass the gem
and parse the response to open struct, so everything on the
client side will work as usual.

Fixes #131