Fix unknown file upload related issues
Currently, the ribose gem is strictly looking for a file type so
if there is no extension exists then it's throwing an error, which
is not expected. The ribose app is also allowing user to upload an
unknown type of file.

So, this commit changes the existing behavior and it should allow
user to upload any type of file, and when it can't determine the
file type then it will sent `application/octet-stream` as type

Fixes #130