Add support to download a file version
This commit adds an interface, that will allow user to download
any file version from a user space. The required arguments for
this interface is `space_id` and `file_id`, and if nothing else
is provided then it will do additional API request to fetch the
remaining details. Usages:

```ruby
Ribose::SpaceFile.download(
  space_id, file_id, version_id: 1, output: "output_file", **options,
)
```