Adopt changes for new auth and ruby 3.0
The new Ribose API require us to provide some more details with each
of the request to comply with device token based authentication. This
includes having a `uid`, `client` and short lived `access-token`.

This commit adopt those changes and add support to handle those API
details under the hood. So, the main changes from a user's end will
be providing `user_email` and `password`, and this library will handle
the rest.

This commit also add support for ruby 3.0, so now we should be able
to run this gem on ruby `2.5` or later.

Fixes #140